requirements.txt
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
streamlit>=1.33.0
python-pptx>=0.6.23
requests>=2.31.0
faiss-cpu>=1.8.0
sentence-transformers>=2.7.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
